+++ nitrite-fjall-adapter/src/lib.rs
@@ -16,7 +16,7 @@ pub use module::*;
 mod tests {
     use crate::map::FjallMap;
     use crate::store::FjallStore;
-    use fjall::{TxKeyspace, TxPartitionHandle};
+    use fjall::{SingleWriterTxDatabase as TxKeyspace, SingleWriterTxKeyspace as TxPartitionHandle};
     use nitrite::common::NitritePluginProvider;
     use nitrite::store::NitriteMapProvider;